requirements.txt
numpy==1.23.5
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.1
imbalanced-learn==0.11.0
tensorflow-macos==2.13.0
torch==2.0.1
SALib==1.4.7
